Fixed the networking features of SmartPM
The files "Constants.java", "EditorPopupMenu.java" and
"EditorToolBar.java" modify the GUI and provide the process designer
with the chance to simulate (with the off-line simulator) or to execute
(with the task-list handler) the process. The file
"DefaultPMProtocol.java" has been customized to send/receive Indigolog
commands to/from the Task-list Handler. Few modifications have been done
to the files "Constants.java" and "RunProcess.java" as well.